+++ backend/src/main/java/woorifisa/goodfriends/backend/common/BaseCreateTimeEntity.java
@@ -1,9 +1,6 @@
-package woorifisa.goodfriends.backend.global.common;
+package woorifisa.goodfriends.backend.common;
 
-
-import lombok.experimental.SuperBuilder;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import javax.persistence.Column;
@@ -13,24 +10,16 @@ import java.time.LocalDateTime;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class) // 3, 4
-public abstract class BaseTimeEntity {
+public abstract class BaseCreateTimeEntity {
 
     @CreatedDate // 1
     @Column(name = "created_at", nullable = false, updatable = false)
     private LocalDateTime createdAt;
 
-    @LastModifiedDate // 1
-    @Column(name = "last_modified_at", nullable = false)
-    private LocalDateTime lastModifiedAt;
-
-    public BaseTimeEntity() {
+    public BaseCreateTimeEntity() {
     }
 
-    public LocalDateTime getCreatedAt() {
+    public LocalDateTime getCreatedDate() {
         return createdAt;
-    }
-
-    public LocalDateTime getLastModifiedAt() {
-        return lastModifiedAt;
     }
 }
