+++ backend/src/main/java/woorifisa/goodfriends/backend/offender/domain/Offender.java
@@ -1,14 +1,10 @@
 package woorifisa.goodfriends.backend.offender.domain;
 
 import lombok.Builder;
-import lombok.experimental.SuperBuilder;
-import org.hibernate.annotations.OnDelete;
-import org.hibernate.annotations.OnDeleteAction;
-import woorifisa.goodfriends.backend.global.common.BaseCreateTimeEntity;
+import woorifisa.goodfriends.backend.common.BaseCreateTimeEntity;
 import woorifisa.goodfriends.backend.user.domain.User;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
 @Table(name = "offenders") // 부정행위자
 @Entity
